app/requirements.txt
flask
SQLAlchemy==1.4
flask-restx==0.5.1
pinecone
openai
python-dotenv
gunicorn==21.2.0
supabase